app/requirements.txt
pandas
awsiotsdk
awscrt